workers/constructors/SharedWorker/unexpected-global-properties.html

<!doctype html>
<title>unexpected members/interface objects/constructors</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  const worker = new SharedWorker('unexpected-global-properties.js');
  worker.port.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data, '');
  });
}, 'Test unexpected properties are not in global scope');
</script>
